requirements.txt
discord-py-interactions
python-dotenv
spotipy
pytz
pandas
requests
python-dotenv
prettytable
Pillow
rdoclient
pymongo
asyncssh
mcstatus
pykuma
notion-client
twitchAPI
babel
isodate